+++ Sources/Extensions/SwiftStdlib/Deprecated/SwiftStdlibDeprecated.swift
@@ -5,13 +5,48 @@
 //  Copyright © 2016 SwifterSwift
 //
 
+// MARK: - Properties
+public extension Bool {
+
+    /// SwifterSwift: Return inversed value of bool.
+    ///
+    ///        false.toggled -> true
+    ///        true.toggled -> false
+    ///
+    @available(*, deprecated: 4.3, message: "Use !self instead.")
+    public var toggled: Bool {
+        return !self
+    }
+
+}
+
+// MARK: - Methods
+public extension Bool {
+
+    @discardableResult
+    /// SwifterSwift: Toggle value for bool.
+    ///
+    ///        var bool = false
+    ///        bool.toggle()
+    ///        print(bool) -> true
+    ///
+    /// - Returns: inversed value of bool.
+    @available(*, deprecated: 4.3, message: "Use !self instead.")
+    public mutating func toggle() -> Bool {
+        self = !self
+        return self
+    }
+
+}
+
+
 extension String {
 
 	/// SwifterSwift: Number of characters in string.
 	///
 	///		"Hello world!".length -> 12
 	///
-	@available(*, deprecated: 4.1.0, message: "This extension is deprecated, please use .count instead.")
+	@available(*, deprecated: 4.1.0, message: "Use .count instead.")
 	public var length: Int {
 		return count
 	}
@@ -23,7 +58,7 @@ extension String {
 	///
 	/// - Parameter i: string index the slicing should start from.
 	/// - Returns: sliced substring starting from start index (if applicable) (example: "Hello world".slicing(at: 6) -> "world")
-	@available(*, deprecated: 4.1.0, message: "Please use string[safe: i] instead.")
+	@available(*, deprecated: 4.1.0, message: "Use string[safe: i] instead.")
 	public func slicing(at i: Int) -> String? {
 		guard i < count else {
 			return nil
@@ -40,7 +75,7 @@ extension String {
 	///   - start: string index the slicing should start from.
 	///   - end: string index the slicing should end at.
 	/// - Returns: sliced substring starting from start index, and ends at end index (if applicable) (example: "Hello World".slicing(from: 6, to: 11) -> "World")
-	@available(*, deprecated: 4.1.0, message: "Please use string[safe: start..<end] instead.")
+	@available(*, deprecated: 4.1.0, message: "Use string[safe: start..<end] instead.")
 	public func slicing(from start: Int, to end: Int) -> String? {
 		guard end >= start else {
 			return nil
@@ -55,7 +90,7 @@ extension String {
 	///
 	/// - Parameter string: substring to search for.
 	/// - Returns: first index of substring in string (if applicable).
-	@available(*, deprecated: 4.1.0, message: "Please use string.index(of: Character) or string.range(of: StringProtocol) instead.")
+	@available(*, deprecated: 4.1.0, message: "Use string.index(of: Character) or string.range(of: StringProtocol) instead.")
 	public func firstIndex(of string: String) -> Int? {
 		return map({ String($0) }).index(of: string)
 	}
@@ -68,7 +103,7 @@ extension String {
 	///
 	/// - Parameter separator: separator to split string by.
 	/// - Returns: array of strings separated by given string.
-	@available(*, deprecated: 4.1.0, message: "Please use string.split(separator: Character) instead.")
+	@available(*, deprecated: 4.1.0, message: "Use string.split(separator: Character) instead.")
 	public func splitted(by separator: Character) -> [String] {
 		return split { $0 == separator }.map(String.init)
 	}